Add files via upload
I fixed the possibility of hearing (and seeing) the input signal or the output signal. Awesome! In particular I made it so that (thanks to a simple if) the various transfer functions calculated the output according to the listening selection made: if the user chooses to listen to the ouput the transfer function is calculated as before, if he chooses the input instead the distortion calculation part is bypassed. I made sure that changing the values of the four knobs only has an effect if I am listening to the output, otherwise changing those values has no effect on the signal. The only exception is the input gain knob, which I have slightly modified so that, if I am listening to the input, I can modify the sound, but only up to 1; going above this value would still risk distorting the signal, even if I don't want it to (this is clear with the sine wave, which already has amplitude 1).